(maint) Add max_connections parameter
Previous to this commit, the PSQL server setting 'max_connections' was
not managed and instead left to the user to tune. However, starting with
PE 2016.1.1, this value needs to be a minimum of 200 in order for PE to
work, as the default PE stack will consume more than 100 connections at
startup.
This commit adds a parameter to change it, with a default of 200.