Default postgresql version based on PE version
Prior to this commit, we defaulted to postgresql 9.4 always.

After this commit, if no version is explicitely given then we use
9.6 but if the class is being applied via the agent running on the
external postgresql node then we will determine the version to use
based on the version of the master.  This will allow users on
versions older than 2017.3 to upgrade the module if they wish.

However, I don't think the module will upgrade postgresql from
9.4 to 9.6 for users.